luminary/pyproject.toml
[project]
name = "luminary"
version = "0.2.0"
description = "Luminary — FastAPI Enterprise API Explorer"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiofiles>=24.1.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "respx>=0.21.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
